+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ConfigShellCommandViewModel.kt
@@ -1,17 +1,18 @@
 package io.github.sds100.keymapper.base.actions
 
 import android.os.Build
-import android.util.Base64
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import io.github.sds100.keymapper.base.R
 import io.github.sds100.keymapper.base.utils.ProModeStatus
 import io.github.sds100.keymapper.base.utils.navigation.NavDestination
 import io.github.sds100.keymapper.base.utils.navigation.NavigationProvider
 import io.github.sds100.keymapper.base.utils.navigation.navigate
+import io.github.sds100.keymapper.base.utils.ui.ResourceProvider
 import io.github.sds100.keymapper.common.models.ShellExecutionMode
 import io.github.sds100.keymapper.common.models.isExecuting
 import io.github.sds100.keymapper.common.utils.Constants
@@ -20,6 +21,7 @@ import io.github.sds100.keymapper.data.Keys
 import io.github.sds100.keymapper.data.repositories.PreferenceRepository
 import io.github.sds100.keymapper.sysbridge.manager.SystemBridgeConnectionManager
 import io.github.sds100.keymapper.sysbridge.manager.SystemBridgeConnectionState
+import java.util.Base64
 import javax.inject.Inject
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.map
@@ -32,7 +34,9 @@ class ConfigShellCommandViewModel @Inject constructor(
     private val navigationProvider: NavigationProvider,
     private val systemBridgeConnectionManager: SystemBridgeConnectionManager,
     private val preferenceRepository: PreferenceRepository,
-) : ViewModel() {
+    resourceProvider: ResourceProvider,
+) : ViewModel(),
+    ResourceProvider by resourceProvider {
 
     var state: ShellCommandActionState by mutableStateOf(ShellCommandActionState())
         private set
@@ -68,11 +72,11 @@ class ConfigShellCommandViewModel @Inject constructor(
     }
 
     fun onDescriptionChanged(newDescription: String) {
-        state = state.copy(description = newDescription)
+        state = state.copy(description = newDescription, descriptionError = null)
     }
 
     fun onCommandChanged(newCommand: String) {
-        state = state.copy(command = newCommand)
+        state = state.copy(command = newCommand, commandError = null)
         saveScriptText(newCommand)
     }
 
@@ -84,8 +88,14 @@ class ConfigShellCommandViewModel @Inject constructor(
         state = state.copy(timeoutSeconds = newTimeoutSeconds)
     }
 
-    fun onTestClick() {
+    fun onTestClick(): Boolean {
         testJob?.cancel()
+
+        val commandError = validateCommand(state.command)
+        if (commandError != null) {
+            state = state.copy(commandError = commandError)
+            return false
+        }
 
         state = state.copy(
             isRunning = true,
@@ -95,6 +105,8 @@ class ConfigShellCommandViewModel @Inject constructor(
         testJob = viewModelScope.launch {
             testCommand()
         }
+
+        return true
     }
 
     private suspend fun testCommand() {
@@ -119,7 +131,21 @@ class ConfigShellCommandViewModel @Inject constructor(
         )
     }
 
-    fun onDoneClick() {
+    fun onDoneClick(): Boolean {
+        val commandError = validateCommand(state.command)
+        if (commandError != null) {
+            state = state.copy(commandError = commandError)
+            return false
+        }
+
+        if (state.description.isBlank()) {
+            state = state.copy(
+                descriptionError = getString(R.string.error_cant_be_empty),
+            )
+
+            return false
+        }
+
         val action = ActionData.ShellCommand(
             description = state.description,
             command = state.command,
@@ -133,6 +159,8 @@ class ConfigShellCommandViewModel @Inject constructor(
         viewModelScope.launch {
             navigationProvider.popBackStackWithResult(Json.encodeToString(action))
         }
+
+        return true
     }
 
     fun onCancelClick() {
@@ -150,9 +178,24 @@ class ConfigShellCommandViewModel @Inject constructor(
         }
     }
 
+    /**
+     * @return the error message.
+     */
+    private fun validateCommand(command: String): String? {
+        if (state.command.isBlank()) {
+            return getString(R.string.action_shell_command_command_empty_error)
+        }
+
+        if (state.command.trimStart().startsWith("adb shell")) {
+            return getString(R.string.action_shell_command_adb_shell_error)
+        }
+
+        return null
+    }
+
     private fun saveScriptText(scriptText: String) {
         viewModelScope.launch {
-            val encodedText = Base64.encodeToString(scriptText.toByteArray(), Base64.DEFAULT).trim()
+            val encodedText = Base64.getEncoder().encodeToString(scriptText.toByteArray()).trim()
             preferenceRepository.set(Keys.shellCommandScriptText, encodedText)
         }
     }
@@ -162,7 +205,7 @@ class ConfigShellCommandViewModel @Inject constructor(
             preferenceRepository.get(Keys.shellCommandScriptText).collect { savedScriptText ->
                 if (savedScriptText != null && state.command.isEmpty()) {
                     try {
-                        val decodedText = String(Base64.decode(savedScriptText, Base64.DEFAULT))
+                        val decodedText = String(Base64.getDecoder().decode(savedScriptText))
                         state = state.copy(command = decodedText)
                     } catch (e: Exception) {
                         // If decoding fails, ignore the saved text